feat: dropdown for currency properties
