+++ packages/nc-gui/package.json
@@ -26,6 +26,7 @@
     "httpsnippet": "^2.0.0",
     "inflection": "^1.12.0",
     "jsep": "^1.3.6",
+    "locale-codes": "^1.3.1",
     "material-design-icons-iconfont": "^5.0.1",
     "monaco-editor": "^0.19.3",
     "monaco-themes": "^0.2.5",
